Correct pagination issue (2)

+++ db/generic.go
@@ -265,7 +265,7 @@ func (dl Datalist) Slice(limit, offset int64) Datalist {
 	if offset >= 0 {
 		dl = dl[offset:]
 	}
-	if limit < len(dl) {
+	if limit < int64(len(dl)) {
 		dl = dl[:limit]
 	}
 	return dl
